+++ src/main/java/com/postgraduate/domain/mentoring/application/usecase/MentoringManageUseCase.java
@@ -5,16 +5,12 @@ import com.postgraduate.domain.account.domain.service.AccountGetService;
 import com.postgraduate.domain.mentoring.application.dto.req.MentoringApplyRequest;
 import com.postgraduate.domain.mentoring.application.dto.req.MentoringDateRequest;
 import com.postgraduate.domain.mentoring.application.dto.res.ApplyingResponse;
-import com.postgraduate.domain.mentoring.application.mapper.MentoringMapper;
 import com.postgraduate.domain.mentoring.domain.entity.Mentoring;
 import com.postgraduate.domain.mentoring.domain.service.MentoringGetService;
-import com.postgraduate.domain.mentoring.domain.service.MentoringSaveService;
 import com.postgraduate.domain.mentoring.domain.service.MentoringUpdateService;
-import com.postgraduate.domain.mentoring.exception.MentoringDateException;
 import com.postgraduate.domain.mentoring.exception.MentoringPresentException;
 import com.postgraduate.domain.payment.application.usecase.PaymentManageUseCase;
 import com.postgraduate.domain.payment.domain.entity.Payment;
-import com.postgraduate.domain.payment.domain.service.PaymentGetService;
 import com.postgraduate.domain.payment.exception.PaymentNotFoundException;
 import com.postgraduate.domain.refuse.application.dto.req.MentoringRefuseRequest;
 import com.postgraduate.domain.refuse.domain.entity.Refuse;
@@ -51,24 +47,13 @@ public class MentoringManageUseCase {
     private final SalaryGetService salaryGetService;
     private final SalaryUpdateService salaryUpdateService;
     private final PaymentManageUseCase paymentManageUseCase;
-    private final MentoringSaveService mentoringSaveService;
-    private final PaymentGetService paymentGetService;
     private final SlackErrorMessage slackErrorMessage;
+    private final MentoringApplyingUseCase mentoringApplyingUseCase;
 
-    @Transactional
-    public ApplyingResponse applyMentoringWithPayment(User user, MentoringApplyRequest request) {
-        Payment payment = null;
+    public ApplyingResponse applyMentoring(User user, MentoringApplyRequest request) {
         try {
-            payment = paymentGetService.byUserAndOrderId(user, request.orderId());
-            mentoringGetService.checkByPayment(payment);
-            String[] dates = request.date().split(",");
-            if (dates.length != 3)
-                throw new MentoringDateException();
-            Senior senior = payment.getSenior();
-            Mentoring mentoring = MentoringMapper.mapToMentoring(user, senior, payment, request);
-            mentoringSaveService.save(mentoring);
-            Optional<Account> account = accountGetService.bySenior(senior);
-            return new ApplyingResponse(account.isPresent());
+            ApplyingResponse applyingResponse = mentoringApplyingUseCase.applyMentoringWithPayment(user, request);
+            return applyingResponse;
         } catch (PaymentNotFoundException ex) {
             log.error("결제건을 찾을 수 없습니다.");
             throw ex;
@@ -76,7 +61,7 @@ public class MentoringManageUseCase {
             log.error("이미 신청된 결제건 입니다.");
             throw ex;
         } catch (Exception ex) {
-            paymentManageUseCase.refundPayByUser(user, payment.getOrderId());
+            paymentManageUseCase.refundPayByUser(user, request.orderId());
             throw ex;
         }
     }
